url logging functionality moved from cumulusurllog to log package
